Docket-system-backend/requirements.txt
Flask
mysql-connector-python
python-dotenv
passlib
PyJWT
flask-cors
